+++ src/vs/editor/contrib/suggest/test/suggestModel.test.ts
@@ -650,7 +650,7 @@ suite('SuggestModel - TriggerAndCancelOracle', function () {
 		return withOracle(async (sugget, editor) => {
 			class TestCtrl extends SuggestController {
 				_onDidSelectItem(item: ISelectedSuggestion) {
-					super._onDidSelectItem(item);
+					super._onDidSelectItem(item, false);
 				}
 			}
 			const ctrl = <TestCtrl>editor.registerAndInstantiateContribution(TestCtrl);
